Listen can be called multiple times. #53.
Also working towards the new API of returning the cancel/stop function
instead of passing back in the callback. This is because you might want
to stop/cancel when an external event fires, and so it's not
convenient to receive that function in the callback. Much better if it's
already in the scope above.