Big push for consistency. #53.
We throw an object with an `error` field now instead of sometimes
calling the callback and sometimes throwing.
We never do one line `if` statements anymore because when the line gets
long we're tempted to just put the body in the line right under, making
us vulnerable to all of the bugs related to forgetting that. Instead
of having a rule for the length at which we use curly brackets and all
we just always do.

`stickerId` renamed `stickerID`.