Update NPM package for getThreadInfo
I was wondering why getThreadInfo wasn't working and it was because it's a recently added feature that hasn't been published to NPM yet. I've updated the version number to reflect the changes (minor update since features were added). Just run `npm publish` and you should be good!

(I'm not sure if you were purposely holding out on publishing or not - if so, I can always just find another way to get thread information)